Move useUpdateCartLocationWhenPaymentMethodChanges to own file
