+++ client/my-sites/checkout/composite-checkout/hooks/use-update-cart-location-when-payment-method-changes.js
@@ -1,0 +1,21 @@
+/**
+ * External dependencies
+ */
+import { useEffect, useRef } from 'react';
+
+export default function useUpdateCartLocationWhenPaymentMethodChanges(
+	activePaymentMethod,
+	updateCartContactDetails
+) {
+	const previousPaymentMethodId = useRef();
+	const hasInitialized = useRef( false );
+	useEffect( () => {
+		if ( activePaymentMethod?.id && activePaymentMethod.id !== previousPaymentMethodId.current ) {
+			previousPaymentMethodId.current = activePaymentMethod.id;
+			if ( hasInitialized.current ) {
+				updateCartContactDetails();
+			}
+			hasInitialized.current = true;
+		}
+	}, [ activePaymentMethod, updateCartContactDetails ] );
+}
